+++ src/Stratis.SmartContracts.Executor.Reflection/ICallDataSerializer.cs
@@ -1,0 +1,12 @@
+﻿using CSharpFunctionalExtensions;
+
+namespace Stratis.SmartContracts.Executor.Reflection
+{
+    /// <summary>
+    /// Serializer for call data sent with a transaction
+    /// </summary>
+    public interface ICallDataSerializer
+    {
+        Result<CallData> Deserialize(byte[] callData);
+    }
+}
