+++ src/Stratis.SmartContracts.Executor.Reflection/ISmartContractVirtualMachine.cs
@@ -1,16 +1,7 @@
-﻿using CSharpFunctionalExtensions;
-using Stratis.SmartContracts.Core.State;
+﻿using Stratis.SmartContracts.Core.State;
 
 namespace Stratis.SmartContracts.Executor.Reflection
 {
-    /// <summary>
-    /// Serializer for call data sent with a transaction
-    /// </summary>
-    public interface ICallDataSerializer
-    {
-        Result<CallData> Deserialize(byte[] callData);
-    }
-
     public interface ISmartContractVirtualMachine
     {
         VmExecutionResult Create(byte[] contractCode,
